+++ app/src/main/java/com/example/sealed_secondhand/recyclerview/PostListRecyclerAdapter.kt
@@ -83,40 +83,40 @@ class PostListRecyclerAdapter(mainActivity: MainActivity) : ListAdapter<PostList
 
         holder.postListLayout.setOnClickListener {
 
-            FirebaseDatabase.getInstance().reference.child("chatRooms")
-                .addValueEventListener(object : ValueEventListener {
-                    override fun onDataChange(snapshot: DataSnapshot) {
-
-                        var isChatRoomExist = false
-
-                        for (childSnapshot in snapshot.children) {
-                            var chatListData: ChatModel = childSnapshot.getValue<ChatModel>() as ChatModel
-                            if (chatListData.postId == getItem(position).postId && (chatListData.destUid == getItem(position).user.firebaseUserId || chatListData.myUid == getItem(position).user.firebaseUserId)) {
-                                isChatRoomExist = true
-                                break
-                            }
-                        }
-
-                        if (!isChatRoomExist) {
-                            var newChatListData = ChatModel()
-                            newChatListData.myUid = FirebaseAuthentication.getCurrentUser()
-                            newChatListData.destUid = getItem(position).user.firebaseUserId
-                            newChatListData.chatTitle = getItem(position).postTitle
-                            newChatListData.postId = getItem(position).postId
-                            newChatListData.lastChat = ""
-                            newChatListData.itemImageUrl = getItem(position).itemImageUrl
-
-                            var data = FirebaseAuthentication.getChatListRoot().push()
-                            data.setValue(newChatListData)
-                        }
-                    }
-
-                    override fun onCancelled(error: DatabaseError) {
-                        // 에러 처리
-                    }
-                })
+//            FirebaseDatabase.getInstance().reference.child("chatRooms")
+//                .addValueEventListener(object : ValueEventListener {
+//                    override fun onDataChange(snapshot: DataSnapshot) {
+//
+//                        var isChatRoomExist = false
+//
+//                        for (childSnapshot in snapshot.children) {
+//                            var chatListData: ChatModel = childSnapshot.getValue<ChatModel>() as ChatModel
+//                            if (chatListData.postId == getItem(position).postId && (chatListData.destUid == getItem(position).user.firebaseUserId || chatListData.myUid == getItem(position).user.firebaseUserId)) {
+//                                isChatRoomExist = true
+//                                break
+//                            }
+//                        }
+//
+//                        if (!isChatRoomExist) {
+//                            var newChatListData = ChatModel()
+//                            newChatListData.myUid = FirebaseAuthentication.getCurrentUser()
+//                            newChatListData.destUid = getItem(position).user.firebaseUserId
+//                            newChatListData.chatTitle = getItem(position).postTitle
+//                            newChatListData.postId = getItem(position).postId
+//                            newChatListData.lastChat = ""
+//                            newChatListData.itemImageUrl = getItem(position).itemImageUrl
+//
+//                            var data = FirebaseAuthentication.getChatListRoot().push()
+//                            data.setValue(newChatListData)
+//                        }
+//                    }
+//
+//                    override fun onCancelled(error: DatabaseError) {
+//                        // 에러 처리
+//                    }
+//                })
             // TODO: PostActivity 추가되면 해당 코드 주석 제거
-            mainActivity.replaceFragment(PostActivity(mainActivity, getItem(position), false))
+            mainActivity.replaceFragment(PostActivity(mainActivity, getItem(position).itemImageUrl, getItem(position), false))
         }
     }
 }
